Create README.md
Business Need:

Predict the future oil prices to make informed decisions regarding buying and selling of oil.

Solution:

To solve this problem, a data science technique was used which involves the use of historical data of oil prices. The data was first loaded into a dataframe using pandas library. The data was then checked for duplicates and missing values. The date column was converted to datetime format and the year column was extracted from the date column. The data was then visualized using different plots to understand the distribution of the prices. The data was then analyzed to find the presence of outliers and the outliers were eliminated using two methods (1) by excluding years that showed outliers and (2) by excluding records with z-score greater than 2.5. The data was then split into train and test datasets and the data was scaled using MinMaxScaler. Finally, a Long Short-Term Memory (LSTM) model was built to predict the oil prices.

Conclusion:

By using data science techniques, the oil price prediction model was built which helped the business to make informed decisions regarding buying and selling of oil. The model was able to make accurate predictions based on the historical data and help the business make profitable decisions.